[cc2] Remove references to MEM
At this moment we have only one enum for everything, and we are using the
same character for the input language, for node operation and for symbol
type. This is bullshit. This patch doesn't solve it, but it remove MEM
that was not used. I have to fix this mesh as fast as possible.