[cc1] Fix parsing of function alike macros
A macro is a function alike macro if just after the name of
the macro there is a '('. Before this commit we were detecting
this situation only looking for tokens, but in this case we
need to check directly in the input line and see if '(' is
not separated from the macro name.