[cc1] Remove TSIZE type
This type was used like a host dependant type which represents
the size quantities, but it generated some problems, mainly
in the printf strings and in the range comparisions (in the
first because we didn't know the correct format specifier and
in the second because we didn't know the maximum/minimun values
of the type). This modification is a step to make cc1 independent
of the host architecture.