Remove MAKEFLAGS use in Makefiles
Per POSIX, "Macros are not exported to the environment of commands to be
run. This was never the case in any historical make and would have
serious consequences. The environment is the same as the environment to
make except that MAKEFLAGS and macros defined on the make command line
are added.", so there is no need to use MAKEFLAGS at all in these cases.

Signed-off-by: Lucas Gabriel Vuotto <l.vuotto92@gmail.com>