[cc2-qbe] Move preambule to code instructions
Until this moment the preambule was writing directly using
printfs calls, but it created some problems related to the
restrictions impossed by qbe about basic blocks. It was
really hard to detect the cases where we were forced
to emit labels to define the begin of a basic block.
With this new approach the preamble is a set of instructions
and we can define where we need labels using the function
introduced a few commits ago.