[cc1] Call decay() in postfix()
Calling decay() in postfix avoids calling it across a lot of different functions
in expr.c, but it has the problem that it makes a bit harder don't decay in the case
of sizeof. This patch is a good compromise between them, because it adds a parameter
to unary() and postfix() and callers set it to the correct value to avoid decaying.