[cc2] Do not use OSTRING in constant()
The OXXXX values are intended for internal use of cc2, and
we do not want to have a dependency between these values
and the intermediate language used between cc1 and cc2.
In this case since it is a language used in both cases
and well documented is more significative to use the actual
value than some stupid macro.