[tests] Fix 0041-queen.c
The function Go was of type int, but it was not returning anything,
although the returned value was not used for anything, so the
behaviour of this function is totally defined. At this moment
QBE does not support returning from a function that return something
without returning something, and since it is not an important point
for this test, it is better to add an implicit 0.