[cc1] Reduce the number of items in arch.h
Arch.h must dissapear because it is the responsible of having
different cc1, while it is not strictly needed. The worst part
is that we need different version of the same object files,
and it is generating a lot of problems in the compilation
process. This is only a first step to remove this problem.