Add sccsid strings
Due to the last changes in the makefile the build process is a bit
unstable now, and in some cases we think that is possible to link
objects of different architectures. Using sccsid strings and the
what command can help us to detect problems.

Be aware that gcc is too clever and knows better than you what
you want to do, and if it has some level of optimization it
deletes the strings if there is some level of optiization,
so if you want to see them use -O0.