[cc1] Rewrite the input system
The input system was broken for a lot of different reasons. Some of them
were due to the fact that the responsability of the different functions
was not clear, and the kind of input wasn't explicit. This new version
tries to split clearly the responsability of every function (for
example spaces are skiped only in skipspaces now).