[cc1] Add type field to input structure
The type is implicit in several of the other fields
of the sructure but it is generating several problems,
because guessing continuosly the type is not easy always.
The number of Input structures is going to be small, so
the space wasted is totally insignificant and the
quality of the code can be improved a lot.