Fix namespace for previously declared nested struct identifier
In structdcl, newtag is called which looks up the struct identifier in
NS_TAG, then sets namespace to NS_IDEN and reads the next token.

This is incorrect since if the struct was already declared, the next
token is an identifier, which is looked up in the wrong namespace.

This causes the following to be incorrectly rejected

	struct S1 { int x; };
	struct S2 { struct S1 s1; };
	struct S2 s2;
	s2.s1.x = 1;

and this to be incorrectly accepted

	struct S1 { int x; };
	struct S2 { struct S1 s1; };
	s1.x = 1;

To resolve this, make newtag preserve namespace. This also allows slight
simplification of structdcl.