Only add to localtypes if curfun is set
Otherwise, a function declaration in the parameters of another function
will by added to localtypes. If the outermost function is just a
declaration with no definition, the type will remain in localtypes after
the declaration. After the next function is defined, flushtypes will
try to remove the function parameter type and assume that it is at the
front of the type table, but other types may have been declared in the
global context since then.

This fixes an assertion failure when HASH(t) is defined as a constant
for

  void f(void (*g)(void));
  int main() { return 0; }