[cc1] Fix install()
Install() was only checking if the symbol to install was already
defined before creating a new symbol, but it fails in the case of
functions, where you can have a parameter with the same name than
the function. In this case the parameter is installed before the
function symbol is installed.