[tests] Move cc1/tests to tests/
All the tests in cc1/tests are from a legacy test suite that is not
used anymore. This test tested cc1 IR, and it was very good to test
internal aspects of cc1, but for this reason it was very difficult
to maintain it accurate, because any small change in cc1 generated
a fail in the tests. Tests/ is a more convenient test framework
since it may be used for any architecture and not only for z80.