Chage makefile framework
At this moment there were several problems with the makefiles,
because we only could configure the archs and systems that
we want in the final build, and it meant that the final
target list was a cartesian product of both. This set
included non desired (and absurd) combinations like for example
z80-linux. Other problem was the impossibility of having
support for the new qbe support for aarch64.

This new framework uses a dep rule which generates makefiles
(makefiles are different files than Makefiles), and it helps
to avoid all the for loops that were needed for generating
all the requested architectures. All the needed information
is now in the leaf makefiles.