[cc1] Add warning about invalid input character
Ctype functions have undefined behaviour out of the range of unsigned char,
and they are only meanful in the character set defined by the current locale,
that in our case is the "C" locale. The "C" locale warranties that all the
characters are representable in the range of positive chars.

The value returned by getchar() is always a number in the range of unsigned
char or EOF, so the use of ctype functions is safe in readchar() and it
will generate the expected behaviour. After that point (and after showing
the proper warning) the value is stored in a char array, and using that
value from the char array will produce a sign extension if char is signed
and the value is negative, and the resultant value will be an invalid value
for ctype functions, and the result will depend of the implementation of
the libraries used to compile scc. The shame of UB is yours!!!