[ar] Add base of -m
M key is harder because in the first pass
we have to generate 2 temporary files instead of only one.
Although it can be with out previous run() function it
could mean that we need and additional step; one for selecting
the files to move, and other to copy the all the other files.