[nm] Improve Makefile
This patch introduces mkdep.sh script which can be used
iin other Makefiles to update the inclusion dependencies.
In our project is so easy because we don't have recursive
inclusion and we always use " for our internal headers.
In the few cases where we use <> then we have to set them
manually.