Fix syntax of jumps
'JP [HL]' and similar imply that the contents pointed by [HL] are read,
and that is the address where the CPU jumps to. This is not the case.
That syntax would mean that a byte is loaded from the address pointed by
HL, and a byte is not a valid address.

The actual behaviour is to move the value of HL to the PC, which means
that there is no indirection, and that the syntax should be 'JP HL'.

The same applies to 'JP [IX]' and 'JP [IY]'.