[libc] Ignore the executable format
At this moment we are not mature enogh to think about the executable
format, and it is creating problems. It can be added back once
that the linker is ready and supporting different file formats.
For now, we are going to assume that every system only supports
one file format.