[scripts] Add mklib script
This script uses lorder to get a correct order in the library
and it uses -q or -ru to make faster the creation of the library.
Some build options of GNU make make make not POSIX compliant
and the libc is fully replaced in the archive every time.
Send a patch to GNU to solve that problem.