[ld] Make undef() private
Undef() was public only because we had to parse the options
in main, but the actions were totally private to the 1st pass.
This patch makes pass1 to rescan the arguments to locate
the options that are interesting for it, and giving a positional
value to -u. -U was evaluated before scanning any file, but ater
this change it is evaluated just where it is found.