libc: Fix malloc() in linux
Linux brk() syscall does not follow the SUSv2 behaviour.
SUSv2 mandates brk() to return an integer that indicates
if the syscall was correct or not, but Linux returns the
previous heap in case of error. Sbrk() in malloc.c   was
trying to solve that problem with a weird comparision,
that generated an integer to pointer conversion, because
the prototype of _brk() was returning integer but   the
actual syscall was returning a pointer, truncating the
value to 32 bits. This patch adds a wrapper over the Linux
syscall and ry to keep the portable code to the SUSv2
interface.