libc: Fix _brk() in Darwin
Darwin lacks a brk() syscall, and the system code was implement
a very basic version using static memory. The code was using the
Linux prototype, while we moved to use the SUSv2 interface, that
just returns 0 or -1 in case of error. The implementation didn't
set the value of errno in case of error, that could generate
strange error diagnosis.