build: Add fixes for NetBSD
NetBSD needs the same hack for ARFLAGS that OpenBSD.
Ideally we should submit a patch for NetBSD to solve
this issue because it has a good support for .POSIX.

This patch also adds a temporary hack to solve a
problem in NetBSD when POSIX_C_SOURCE and _ANSI_SOURCE
are defined at the same time. A patch was submited to
NetBSD and this hack is expected to be removed in a
future.