libc: Fix build in bsd systems
After the commit 383f380a the other BSD systems didn't compile
because the new sigaction.h file was not present on them.
This commit uses the same header that the one used in
OpenBSD that should work but it could not be tested correctly.
At least the compilation is fixed again.