libc: Fix fopen with mode "r"
On all supported platforms, O_RDONLY is 0, so flags & O_RDONLY is
never true and the FILE does not get marked as _IOREAD, causing all
read operations to fail with EBADF.

To fix this, introduce O_ACCMODE in sys.h, containing a mask of the
bits used for the file access mode. Then, in _fpopen, switch on
access mode.