build: Select set of libc targets based in uname
The libc depends of the compiler to be able to generate
code for different processors, and trying to compile a
system without support from the toolchain will make the
compilation fail. In order to try to minimize this point,
this commits uses uname -m to select the set of confi-
gurations that must be build.