build: Split proto files
Until this moment it was hard to install architectures
different than amd64. Having them in the proto file
generated errors when they were not compiled, forcing
to compile all the architectures always. This is a problem
in this stage because we need external toolchains for
the libc, incrementing the list of external dependencies.

This modification splits the common part and the part that
depends of the toolchain used. It only adds the proto for
amd64, that is the most common architecture used by the
scc developers, but it provides scripts to make easy to
update any proto file needed.