build: Add support for DESTDIR
PREFIX is used to mark what is going to be the final
path where the program is going to be stored and where
it will locate its resources. Package creators need a
way to redirect all the installation to a directory
that can be used as base doe a tar.gz file instead
of installing all the files in the prefix destination.