libmach/coff32: Use int for index in coff32getsym()
The prototype of getsym() includes an integer pointer
for the index. Several binary formats use bigger data
types for indexes, but in practical cases having a bigger
value does not work because it would require too much
memory to have the full symbol table in memory. Coff32
was converting the index value into long, because the
index value used by coff32 has long type, but it does
not make sense because it will not increase the range
already limited by the prototype of the function.