scc: Use scc prefix for tools
Scc is in an early stage and it is not stable enough
to be used instead of the systems tools. For this
reason this patch renames all the binaries to have
a scc- prefix that avoid executing the wrong tool at
the wrong moment.

This patch also adds a new tool called scc that sets
automatically the value of SCCPREFIX based in the name
of the program and simplify executing scc from a
directory different than $PREFIX.