cc1: Disable escape() while parsing #define
When the expansion part of a #define was parsed the code
was applying the same escape rules than when it parses
normal C code. The result of this was that strings were
escaped in the definition of the macro, so any embedded
and escaped " would be expanded without the escape sequence.
This change disables escape for strings when a define is
parsed, but cc2-qbe is still passing an unescape string
to qbe.