cc1: Allocate an identifier for local initializers
In the case of local initializers a hidden variable is emitted,
but this variable lacked an identifer. In this case, the default
identifier is 0, that cc2 understand as a temporary value, which
meant that it was not inserted in the hash table. Every access
to the symbol identified by 0 was generating new symbols, without
relation between them.