cc2/qbe: Change ternary() to return Node
Receiving a pointer to the struct were we want to store the
result needs a lot of temporary variables in the code. If
we make that ternary() returns a struct instead of a pointer to
struct we can remove a lot of these temporaries and
simplify the code. The code generated is pretty similar
because returning structs usually mean passing an additional
pointer to the function ternaryed where it can write the results.