cc2/qbe: Simplify field()
The offset check is not needed because qbe can figure out
the optimization. This target is the simplest of all  the
targets because the big part of the problem is solved by
qbe, so it is better to try to keep it as simple as possible
in a way that can be used as skeleton for other targets.