cc2/qbe: Fix post increment and decrement
These operators use their operand as lhs and rhs, and assign()
was evaluating twice the operand. It didn't have effects with
simple variables but it created problems with more complex
expressions with side effects. The operand must be evaluated
only once, like a lhs and load it when it has to be used like
an rhs.