cc1: Fix double free() error
The current code was too complex without a good reason. The lexer was not
trying to concatenate consecutive strings, and this work was done in the
expression parser. This was an error because it introduced complexity due
to the memory allocation needed to concatenate the strings, and the only
reason to do this was to avoid moving lineno variable that could produce
more difficult to read error messages because they could be moved to
different lines. That problem must be solved in a different way, where
the variable line that is used in error reporting must be updated only
when the beginning of a new token is detected.