cc1: Centralize hash table initialization
The hash table was initialialized in several places, and in all
of them the counterid was reset to 0. This was very dangerous
and the types emitted in irach() were using the same id than
the next types and variables declared in the code. To avoid
this problem and to minimize the dependencies we place all the
initialization in only one place located in symbol.c.