libc: Merge&split correctly with next block
When the consecutive block is the next block we can merge it with
the block of realloc'ed pointer. In that case we have to remove
the next block from the free list, and insert the remaining part
of the merge in the list. To remove the next block we have to
modify the next pointer of the previous block and assign the next
block of the next pointer to the next pointer of the new block.