as: Undef BIG_ENDIAN and LITTLE_ENDIAN
There are some systems where BIG_ENDIAN and LITTLE_ENDIAN are defined
like macros, even when a pure c99 environment is requested. We tried
hard to force these systems to solve the problem, but we couldn't get
them fixed in some systems, and in other we didn't find a place where
to send the patches.