scc: Use CONF as default configuration
The driver was using the first element of SYSLST as default configuration
which is not very intuitive if you want to use scc as a system compiler.
This change makes CONF the default configuration and it adds auto detection
in the all target. In case of needing a different default it can be obtained
compiling scc with:

	make CONF=xxxxx toolchain