cc1: Revert commit 7d04db92
Macro parameters must be expanded when they are substitued and
not when they are detected. At this moment we don't have a way
to hide symbols if we expand them in copypars(), so for this
reason is better to mark as TODO the tests that stop working
after this change.