cc1: Use alignment in the offset calculation
The function typesize() was updating the base offset of
every field, but it was not considering the alignment
for the calculation of the offset. The alignment was
still considered for the actual size of the type, and
since code.c was emitting padding bytes when the alignment
was not correct the emited layout was correct, but the
definition of the tyoe was not correct. Instructions
using the offset of the field were not taking in account
the offset added due to the padding bytes. This error
was detected by the test 0050-inits.c.