cc1: Improve readline() readbility
Having readline() and cpp() in the same if was a bit confusing
because it seemed like there was some relation between them, but
they are totally independent. In the same way, at EOF there was
an implicit condition of having an empty string pointed by p
but it was really obsucre how it was happening.