cc1: Add type parameter to addinput()
Addinput() had a complex logic based in the received parameters to
decide what action should be taken. This logic was hard from the
external caller and it created confusing situations, that joined
to the fact that was not clear who was the owner of the memory
made everything a bit complex. This patch solves the problem adding
a type parameter and moving ownership of memory inside addinput()
which is responsible to allocate memory now.