cc1: Move defined function to cpp.c
Defined() enables and disables the macro expansion, and as it is
a very delicated thing to do it is better to keep all the code
that modifies disexpand in only one file.

While this patch was written it was discovered that the function
notdefined() was not acutally needed because the code to handle
preprocessor defined was ran from unary().