libc: Return 0 when truncation in strftime()
When the buffer is not big enough to hold the generated string strftime()
must returns 0. Strftime() was genereting a marker value '?' in that
case, but it made impossible to control the actual space that would be
required to store the value. For that reason, the code is modified
to store as much characters as it can, enabling it to detect the
truncation case.