libc: Pull atexit() buffer only when needed
Exit() has to call all the functions registered with atexit()
and it means that the code was pulling always the full buffer
even when atexit() was not called, that in some architectures
could be a large value. This code encapsulate the loop in a
function pointer that is set in every call to atexit(). This
could be done with common symbols but since there are some
portabilities issues around them we used this other solution
that will increase the size of bss in one pointer when atexit()
is called.