cc1: Fix hideset implementation
The correct macro expansion is based in the hideset concept.
When the macro is expanded and scanned for tokens all the
macros expanded in the context of the macro expansion must
be disabled and not expanded again. It means that the hide
set of every macro expansion cannot be represented only with
a symbol but with a set of symbols.

It is important to hide the symbols after they are expanded
in the context of the macro expansion until the macro
expansion finished.