cc1: Accept empty parameters in macros
A macro accepting a parameter can be instantiated without
anything between the parenthesis, because in that case it
receives an empty parameter. We have to differentiate that
case from the case that acutally has no parameter.

As this patch modifies the behaviour and ahead() is not called
always befor calling parameter() we have to explicit trim
spaces at the beginning of the parameters.