driver/posix: Don't free in error
In case of error we were freeing some memory, but some of that
memory was not dynamically allocated, and in some cases it was
string contants. The best approach in this case is not freeing
anything because we are not going to save anything since at that
stage we can only finish with error.