driver/posix: Add a template for ld execution
The command line for ld is going to depend deeply in the libc used.
As the scope of scc libc is to provide a simple c99 library can be
easily ported to different targets it is not going to support a full
POSIX interface. In order to compile POSIX programs with scc the
best approach is to use musl libc, which has portable headers that
scc can use without problems. Musl has a link command line totally
different to the one used by scc libc and for that reason the
template mechanism is provided.