cc1: Don't remove warnings in typedef'ed functions
When a typedef of a function is used as base to build a new typedef
then it does not have any argument because in that case the typedef
only carries the type but no arguments. The pars field of the decl
structure was not initialized because this case was not expected and
any NULL check was failing.