cc1: Accept () in primary()
Accepting parentheses was moved to cast() because we require a 2 token
ahead, and the only possible way of doing that was accepting parenthesis
there. The problems comes because unary() calls primary() and unary() is
called from sizeexp() that is called from unary(). It means that a sizeof
expression could not begin with 2 parenthesis (one was already removed
in sizeexp()) and for that reason we have to accept parenthesis in
primary().