as: Initialize nvar variable
Non initialized variables in awk have the empty string value,
that when it is used in a numerical context is evaluated to 0,
but it is still evaluated to "" in the context of an array
index. Nvar was used later to run over the arrays but 0
didn't have any content in the arrays because nvar evaluated
to "" in the first iteration.