libc/stdio: Use size_t for cnt in vfprintf()
The cnt variable is used to count the number of bytes written in
every iteration of the loop and it was added to the len variable.
While the len variable was size_t cnt was an int and we were storing
things like size of a string that can have a size_t len. Also,
scc IR has a problem with abbreviations where the result type
is smaller than the original type, and this was one of this cases.