libc/stdio: Simplify string output in vfprintf()
It is not needed to calculate the lenght of a string because
we already check it while we print it and it is cross checked
against the precision. We can use SIZE_MAX instead of -1 because
it works as a positive number, removing several of the problems
that the code had.