libc: Remove mbstate_t handling
We only support UTF-8 that is stateless, so it does not make sense
to keep static variables and handle them when they are not going to
have any effect. We tried to keep this only in the lowest level but
it was generating too much edge cases with low benefit that it didn't
make sense.