build: Add initial support for CONF=amd64-freebsd
This is only the basic skeleton to build amd64 FreeBSD. Additional
work is required to have a working libc. It is only checked that
the build finish correctly but we cannot be sure that all the
constants defined in this commit are correct. The errno.h file was
verified and the _exit() syscall was also checked.