cc1/cc2: Fix abbreviations with reduce cast
When the left part of an abbreviation operation is smaller than
the right part then the code was generating casts in the left
part that by definition is wrong. The fix in cc1 introduces a
new case in cc2 because then we can have operations with
size smaller than 4 bytes, that is not allowed by qbe.