cc2/qbe: For use of .L in local variables
ELF and COFF use .L for local labels that are not
exported while a.out uses L for them. In the current
QBE state is unlikely that it will support targets
where .L is not accepted for it, and there are some
assemblers that reject local labels that begin with
. and a digit. This commit forces to use .L always
so the lables are going to be safe for any qbe target
supported.