cc1: Generate endif errors only in actual EOF
The pushup strategy used in the preprocessor generates
EOF tokens in macro expansions that don't mean the end
of the input stream. When that situation what happening
(a macro expansion with parameters inside of a ifdef)
it generated a '#endif expected', when this should happen
only when no more input is available.