driver/posix: Collapse settool() and inittool()
Inittool() assumes that you only have to add the input name
to the command line, but this is not true in all the tools.
The first step is to remove the separation between inittool()
and settool() that are called always. After this point we
should use some kind of template to build the command line
of every tool.

This commit is known to generate multiple leaks because the string
allocated are never deallocated.