ar: Try to use unlink() instead of remove()
It was known that ar was calling remove() that is not
signal safe from a signal handler. The reason was
that ar is designed as portable code and it cannot
call the signal safe unlink() function. This commit
uses the preprocessor to detect some common macros
defined by UNIX compilers to redirect remove()
to unlink() in that case.