cc1: Allow multiline macro invocations
The code was using the input line to mark the begin and end of the
macro parameter that fails when the macro argument list is split
in several lines.

Switching to an internal buffer in the macro invocation has the
side effect of destroying the whitespaces while we scan for the
argument, that has undesired effects when it is used in combination
with # and ##. For that reason the test 0188-multi_string.c was
disabled because to fix it we have to switch how we run over the
input and record any input instead of just increment the input
pointer. That is a more complex change that requires a different
commit.