make: Fix macro precedence
POSIX mandates the following precedence:

	1- Macros specified on the make utility command line.

	2- Macros defined by the MAKEFLAGS environment variable

	3- The contents of the environment

	4- Macros defined in the inference rules built into make.

It means that macros defined in the makefiles has to read before
macros from MAKEFLAGS and from the command line. The simplest
solution passed by doing a first pass over all the options only
to locate the -f options and read the makefiles. After that
step we can parse the options from MAKEFLAGS and later from the
command line and then we can be sure that the precedence is
always correct.