make: Calculate target stamp in rebuild()
The time stamp of every file was calculate when the target
was created in lookup(), and it was not updated ever. It
lead to keep the old timestamp when the target was checked
in different rules, impling build it several times because
the contidion that fired the first rule kept being  valid
forever (for example when the file didn't exist).