make: Accept character # in command lines
The specification allows the use of the character # in command lines,
allowing comments only in the first column of the line. To allow that
we cannot handle comments in nextline() and we have to move the logic
about comments to the different places where the lines are read,  and
add the correct different handling in every case.