libmach: Add setsec()
This function allows to inject a section in the object
described by the generic section interface. As there is
not a perfect mach between the generic struct and the
object struct then some heuristics have to be applied
to be able to transform between then. Anyway, there will
be cases when the assembler and the loader will have to
uses directly object types.